// ui/src/api/generated/models/usage-health.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UsageHealth
 */
export interface UsageHealth {
    /**
     * 
     * @type {string}
     * @memberof UsageHealth
     */
    'affectedSince'?: string;
    /**
     * 
     * @type {string}
     * @memberof UsageHealth
     */
    'affectedUntil'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UsageHealth
     */
    'available'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UsageHealth
     */
    'complete'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof UsageHealth
     */
    'droppedEvents'?: number;
    /**
     * 
     * @type {number}
     * @memberof UsageHealth
     */
    'incompleteCalls'?: number;
    /**
     * 
     * @type {string}
     * @memberof UsageHealth
     */
    'integrityError'?: string;
    /**
     * 
     * @type {string}
     * @memberof UsageHealth
     */
    'lastWriteErrorAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof UsageHealth
     */
    'migrationError'?: string;
    /**
     * 
     * @type {number}
     * @memberof UsageHealth
     */
    'queueDepth'?: number;
    /**
     * 
     * @type {number}
     * @memberof UsageHealth
     */
    'writeFailures'?: number;
}